Add Tweet Reports endpoint

+++ plato/src/twitter/application/query/get_pending_tweets_query.py
@@ -3,9 +3,9 @@ from plato_cqrs import Query
 
 class GetPendingTweetsQuery(Query):
 
-    def __init__(self, publicationDate: float):
-        self.__publicationDate: float = publicationDate
+    def __init__(self, publicationDate: int):
+        self.__publicationDate: int = publicationDate
 
     @property
-    def publicationDate(self) -> float:
+    def publicationDate(self) -> int:
         return self.__publicationDate
