+++ plato/src/twitter/application/query/get_tweets_by_account_response.py
@@ -11,7 +11,7 @@ class GetTweetsByAccountResponse(QueryResponse):
         return self.__tweets
 
     def appendTweet(self, tweetId: str, accountId: str, description: str,
-                    publicationDate: float, published: bool):
+                    publicationDate: int, published: bool):
         self.__tweets.append({
             "tweetId": tweetId,
             "accountId": accountId,
